requirements.txt
django>=1.8.8
django-taggit>=0.18.0
